Use the appropriate type-cast to keep the 64 bit counter from getting truncated...

+++ arch/arm/cpu/armv7/omap5/timer.c
@@ -48,8 +48,10 @@ int timer_init(void)
 unsigned long long get_ticks(void)
 {
         u32 cvall, cvalh;
+
         asm volatile("mrrc p15, 0, %0, %1, c14" : "=r" (cvall), "=r" (cvalh));
-        return (cvalh << 32) | cvall;
+
+        return (unsigned long long)((unsigned long long)cvalh << 32 | cvall);
 }
 
 /*
